### 2026-02-02 #749 @smwindecker
Pre-submission inquiry: EpiStrainDynamics
Submitting Author Name: Saras Windecker
Submitting Author Github Handle: @smwindecker
Repository: https://github.com/acefa-hubs/EpiStrainDynamics
Submission type: Pre-submission
---

```
Package: EpiStrainDynamics
Title: Infer temporal trends of multiple pathogens
Version: 0.0.1.0000
    c(person("Saras", "Windecker", email = "saras.windecker@gmail.com",
      person("Oliver", "Eales", role = "aut",
      person("James", "McCaw", role = "aut",
      person("Freya", "Shearer", role = "aut",
      person("Milad", "Kharratzadeh", role = "ctb")
    )
Description: 'EpiStrainDynamics' is a statistical framework developed for inferring temporal trends of multiple pathogens from routinely collected surveillance data.
License: Apache License 2.0
Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
Biarch: true
    bayesplot,
    cli,
    Rcpp (>= 0.12.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.18.1),
    rstantools (>= 2.4.0),
    timetk,
    tsibble,
    viridis
    BH (>= 1.66.0),
    Rcpp (>= 0.12.0),
    RcppEigen (>= 0.3.3.3.0),
    RcppParallel (>= 5.0.1),
    rstan (>= 2.18.1),
    StanHeaders (>= 2.18.0)
SystemRequirements: GNU make
    tibbletime,
    units,
    xts,
    zoo
```

EpiStrainDynamics is a bayesian statistical modelling framework capable of inferring case trends of multiple pathogens. Estimating the temporal trends in infectious disease activity is crucial for monitoring disease spread and the impact of interventions.

Yes, for both BS and TS

EpiStrainDynamics is a statistical modelling framework capable of inferring case trends of multiple pathogens. Estimating the temporal trends in infectious disease activity is crucial for monitoring disease spread and the impact of interventions. The target audience for this package is researchers and decision-makers who use surveillance data to monitor disease trends.

The modelling framework presented here has scientific applications related to different surveillance systems covering multiple pathogens (eg. influenza, SARS-CoV-2, dengue), scenarios (seasonal epidemics, non-seasonal epidemics, pandemic emergence), and temporal reporting resolutions (weekly, daily). This methodology is applicable to a wide range of pathogens and surveillance systems.

This software expands upon an existing modelling algorithm, developed carefully for the test case of multiple pathogens modelled in an epidemiological context. There is no other existing relevant software for the earlier versions of these models but there are published manuscripts with sections of stan model code made public.

Eales O, de Oliveira ML, Page AJ, et al. Dynamics of competing SARS-CoV-2 variants during the Omicron epidemic in England. Nat Commun. 2022;13(1):1-11. [10.1038/s41467-022-32096-4](https://doi.org/10.1038/s41467-022-32096-4)

Eales O, Ainslie KEC, Walters CE, et al. Appropriately smoothing prevalence data to inform estimates of growth rate and reproduction number. Epidemics. 2022;40:100604. [10.1016/j.epidem.2022.100604](https://doi.org/10.1016/j.epidem.2022.100604)

NA

Moving over from initial issue https://github.com/ropensci/software-review/issues/738#issuecomment-3708881102

I am attempting to resolve the queries posed by @mpadge in that thread, specifically the size of repo problem.
