### 2026-03-11 #749 @mpadge
@smwindecker It seems that your test suite runs approximately forever. Just running https://github.com/acefa-hubs/EpiStrainDynamics/blob/main/tests/testthat/setup-fitted-models.R in our bot environment  takes around 2 hours, and that's needed before any tests can start. I think the bot just timed out and gave up. You'll see that you've also got timeouts on your [codecov workflows](https://github.com/acefa-hubs/EpiStrainDynamics/actions/workflows/codecov.yaml) and [pkgcheck workflows](https://github.com/acefa-hubs/EpiStrainDynamics/actions/workflows/pkgcheck.yaml). And those time outs are at 6 hours.

We're going to require some runnable tests to proceed here. I can well imagine that the full detail of these `fitted-models` is both useful and necessary for full tests. How about an approach like pre-generating those models and uploading them somewhere so they can be quickly downloaded in tests and used directly there? It's a personal preference, but I can definitely recommend uploading stuff like that as binary objects with a GitHub release, and using rOpenSci's [piggyback package](https://docs.ropensci.org/piggyback) to manage the upload/download processes. You've already got [a package release](https://github.com/acefa-hubs/EpiStrainDynamics/releases/tag/v2.0_ROpenSci-submission) which would be a convenient spot.

I didn't even try running any tests once I had finally generated the `fitted-models`, but since that took about 2 hours on our machine, and the GitHub workflows were still running beyond 6 hours, that suggests a likely need for more speed-ups within the general test suite as well.

There may be other approaches as well, but that would be one workable suggestion. We can discuss potential approaches here, or feel free to open an issue on your repo and ping me there if you prefer. thanks!
