### 2026-03-11 #749 @smwindecker
Thanks Mark. I had revised how these extended tests worked between the last submission and this one (obviously to worse outcome). I will attempt what you're describing.
