### 2026-03-27 #749 @emilyriederer
Thank you @smwindecker -- this looks good to me! Please proceed to a full submission and close this issue

@mpadge -- question for you. Why does the output say BS4.2 is not an actual standard? I see it documented [here](https://stats-devguide.ropensci.org/standards.html#analytic-algorithms)
