### 2026-03-27 #749 @mpadge
> [@mpadge](https://github.com/mpadge) -- question for you. Why does the output say BS4.2 is not an actual standard? I see it documented [here](https://stats-devguide.ropensci.org/standards.html#analytic-algorithms)

Very 🦅 👀 of you @emilyriederer - the [issue linked above](https://github.com/ropensci/statistical-software-review-book/pull/95) has fixed that. That standard was actually incorrectly formatted in our original book. I've now added tests to the book itself to make sure that all standards are correctly formatted.
